requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.6
idna==3.2
numpy==1.21.2
oauthlib==3.1.1
pandas==1.1.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.0
six==1.16.0
tweepy==4.1.0
urllib3==1.26.7
